fix: 요청에 몇몇 값들이 잘못 들어가고 있던 오류 해결

+++ SpikeBeach/Source/SpikeBeach/MultiplaySystem/WebSocketRequests/RoomEnterRequest.cpp
@@ -10,7 +10,7 @@ TArray<uint8> URoomEnterRequest::Serialize()
     Bytes.Append((const uint8*)TCHAR_TO_UTF8(*userAssignedId), userAssignedId.Len() + 1);
     Bytes.Append((const uint8*)TCHAR_TO_UTF8(*token), token.Len() + 1);
     Bytes.Append((const uint8*)TCHAR_TO_UTF8(*clientVersion), clientVersion.Len() + 1);
-    Bytes.Append((const uint8*)&roomId, sizeof(int16));
+    Bytes.Append((const uint8*)&roomId, sizeof(int32));
     return Bytes;
 }
